dashboard/test/acceptance/requirements.txt
git+https://github.com/qaenablers/python-qautils@v1.2.0
PyHamcrest==1.8.3
requests==2.6.0
selenium==2.52.0
